Document running tests
One of the first things I do when checking out a project is to run the
tests. Making it clear how to run them is important to me.

I've changed the ci.yml as it is an example of how to run the tests.

I've set the tests to run with less permissions, because it helps give
confidence that they are safe.

I've switched the docker command to use the container subcommand,
because I think it reads better, and it is the way the docker
documentation recommends.

https://www.docker.com/blog/whats-new-in-docker-1-13/#h.yuluxi90h1om