Eliminate ifs by using bit flag logic
I think the new code is better, because it doesn't repeat the
ServerCapabilities bitflag.

Each of the ServerCapabilities flags have precisely one 1 bit in their
binary representations.

```
if (packet.serverCapabilities & ServerCapabilities.CLIENT_LONG_FLAG) {
  clientParam |= ServerCapabilities.CLIENT_LONG_FLAG;
}
```

The value of the if expression will either be `0` or
`ServerCapabilities.CLIENT_LONG_FLAG`, depending on if
`packet.serverCapabilities` has the same bit set.

Instead we can binary or with the expression directly. If the expression
is 0, then the binary or does nothing. If the expression is
`ServerCapabilities.CLIENT_LONG_FLAG`, then the binary or does the same
thing as before.

The same idea applies for the other 2 ifs.